# review/les-auditor/api.openapi.yaml
openapi: 3.0.3
info:
  title: LES Auditor API
  version: 1.0.0
  description: |
    Manual-entry LES (Leave and Earnings Statement) audit API for military paycheck validation.

    **V1 Scope:** Manual entry only (no PDF uploads/OCR).

    **Key Features:**
    - Auto-fills allowances from official DFAS tables (BAH, BAS, Base Pay, COLA)
    - Users enter actual taxes, deductions, and net pay from their LES
    - Validates FICA = 6.2%, Medicare = 1.45% of taxable gross
    - Validates net pay math (±$1 tolerance)
    - Returns actionable flags (red/yellow/green) with suggestions

  contact:
    name: Garrison Ledger Support
    email: support@garrisonledger.com
servers:
  - url: https://garrisonledger.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
security:
  - ClerkAuth: []
paths:
...
  /les/audit/{id}/clone:
    post:
      summary: Clone Audit for Re-Auditing
      description: Creates a new draft audit with the same data from a previous audit, allowing users to edit and re-run.
      tags:
        - LES Auditor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Audit cloned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  auditId:
                    type: string
                    format: uuid
                    description: ID of the newly cloned audit
                  message:
                    type: string
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session JWT token
...